requirements.txt
Flask
gunicorn
pytest
bert-extractive-summarizer
torch==1.0.1
